+++ src/app/components/game-map/game-map.component.ts
@@ -1,7 +1,8 @@
-import { Component, computed } from '@angular/core';
+import { Component, computed, signal } from '@angular/core';
 import {
   gamestate,
   getWorldNode,
+  setCameraPosition,
   windowHeightTiles,
   windowWidthTiles,
 } from '../../helpers';
@@ -21,6 +22,9 @@ export class GameMapComponent {
     Math.min(gamestate().world.height, windowHeightTiles() + 1),
   );
   public camera = computed(() => gamestate().camera);
+  public isDragging = signal<boolean>(false);
+  public lastPanPosition = signal<{ x: number; y: number } | null>(null);
+  private panSensitivity = 2;
 
   public map = computed(() => {
     const width = this.nodeWidth();
@@ -49,4 +53,42 @@ export class GameMapComponent {
 
     return nodes;
   });
+
+  public panStart(event: MouseEvent): void {
+    event.preventDefault();
+
+    this.isDragging.set(true);
+    this.lastPanPosition.set({ x: event.clientX, y: event.clientY });
+    setCameraPosition(0, 0);
+  }
+
+  public panMove(event: MouseEvent): void {
+    if (!this.isDragging() || !this.lastPanPosition()) {
+      return;
+    }
+
+    const lastPos = this.lastPanPosition()!;
+    const deltaX = event.clientX - lastPos.x;
+    const deltaY = event.clientY - lastPos.y;
+
+    const currentCameraPos = gamestate().camera;
+
+    const newCameraX = currentCameraPos.x - deltaX / this.panSensitivity;
+    const newCameraY = currentCameraPos.y - deltaY / this.panSensitivity;
+
+    setCameraPosition(newCameraX, newCameraY);
+
+    this.lastPanPosition.set({ x: event.clientX, y: event.clientY });
+  }
+
+  public panEnd(): void {
+    this.isDragging.set(false);
+    this.lastPanPosition.set(null);
+  }
+
+  public panLeave(): void {
+    if (this.isDragging()) {
+      this.panEnd();
+    }
+  }
 }
